Create solutions-and-projects
As I will be experimenting with several approaches to legendary-guacamole in several languages it makes sense to create and branch development at init.

As it stands the name legendary-guacamole was generated by GitHub. One thought was to make a web spider and ML project that could find and parse guacamole recipes and perform statistical analysis of ingredients and proportions to the ends of "predicting" the recipe for legendary-guacamole.

For now I am just going to create a collection of useful functions in several languages. Starting with C#.