We are working from a specification
As stated before we are looking to create some useful functionality in multiple languages. Our first idea is to provide a string transformer which would take a string of a certain case type (camelCase, snake_case, etc.) and return a string of another case. In this commit we created some interfaces, delegates, methods, enumerations, and folders. We are tinkering with our approach. We also branched into ideas-and-approaches.